ai-service/requirements.txt
flask==3.0.3
tensorflow-cpu==2.16.1
Pillow==10.4.0
numpy==1.26.4
gunicorn==22.0.0
